Got rid of the stupid short unsigned int's

+++ include/quick.h
@@ -1,11 +1,10 @@
 #ifndef QUICK_H
 #define QUICK_H
 
-void swap(short unsigned int *a, short unsigned int *b);
+void swap(int *a, int *b);
 
-short unsigned int partition(short unsigned int arr[], short unsigned int low, short unsigned int high);
+int partition(int arr[], int low, int high);
 
-void quicksort(short unsigned int arr[], short unsigned int low, 
-short unsigned int high);
+void quicksort(int arr[], int low, int high);
 
 #endif
