+++ include/res.h
@@ -1,8 +1,8 @@
 #ifndef RES_H
 #define RES_H
 
-short unsigned int res_type(int a, int b, int c, int d, int e, int f, int g, int h, int i);
+int res_type(int a, int b, int c, int d, int e, int f, int g, int h, int i);
 
-short unsigned int res_wing(int a, int b, int c, int d, int e, int f, int g, int h, int i);
+int res_wing(int a, int b, int c, int d, int e, int f, int g, int h, int i);
 
 #endif
